+++ src/application/calendar.ts
@@ -1,12 +1,11 @@
-import { HttpClient } from '../http-client/index.js';
-import { Application, ApplicationConfig } from './application.js';
+import { PathFactoryFunction } from '../pco-api-client/types.js';
+import { ApplicationConfig } from './application.js';
 
-export class Calendar extends Application {
-  constructor(
-    client: HttpClient,
-    apiVersion: string,
-    options: ApplicationConfig
-  ) {
-    super(client, apiVersion, 'calendar', options);
-  }
-}
+const paths = new Map<string, PathFactoryFunction>([]);
+
+export const CalendarConfigDefault: ApplicationConfig = {
+  apiVersion: 'v2',
+  appVersion: '2022-07-07',
+  name: 'calendar',
+  paths,
+};