+++ src/application/giving.ts
@@ -1,12 +1,44 @@
-import { HttpClient } from '../http-client/index.js';
-import { Application, ApplicationConfig } from './application.js';
+import {
+  PathFactoryFunction,
+  pathFactoryProvider,
+} from '../pco-api-client/types.js';
+import { ApplicationConfig } from './application.js';
 
-export class Giving extends Application {
-  constructor(
-    client: HttpClient,
-    apiVersion: string,
-    options: ApplicationConfig
-  ) {
-    super(client, apiVersion, 'giving', options);
-  }
-}
+const paths = new Map<string, PathFactoryFunction>([
+  // Root
+  ['giving', pathFactoryProvider('get', '')],
+
+  // Batches
+  ['readBatches', pathFactoryProvider('get', 'batches')],
+  ['createBatch', pathFactoryProvider('post', 'batches')],
+  ['updateBatch', pathFactoryProvider('patch', 'batches/{{ id }}')],
+  ['deleteBatch', pathFactoryProvider('delete', 'batches/{{ id }}')],
+  ['commitBatch', pathFactoryProvider('post', 'batches/{{ id }}/commit')],
+
+  // Batch Groups
+  ['readBatchGroups', pathFactoryProvider('get', 'batch_groups')],
+  ['createBatchGroup', pathFactoryProvider('post', 'batches_groups')],
+  ['updateBatchGroup', pathFactoryProvider('patch', 'batch_groups/{{ id }}')],
+  ['deleteBatchGroup', pathFactoryProvider('delete', 'batch_groups/{{ id }}')],
+  [
+    'commitBatchGroup',
+    pathFactoryProvider('post', 'batch_groups/{{ id }}/commit'),
+  ],
+
+  // Donations
+  ['readDonations', pathFactoryProvider('get', 'donations')],
+  ['createDonation', pathFactoryProvider('post', 'donations')],
+  ['updateDonation', pathFactoryProvider('patch', 'donations/{{ id }}')],
+  ['deleteDonation', pathFactoryProvider('delete', 'donations/{{ id }}')],
+  [
+    'refundDonation',
+    pathFactoryProvider('post', 'donations/{{ id }}/issue_refund'),
+  ],
+]);
+
+export const GivingConfigDefault: ApplicationConfig = {
+  apiVersion: 'v2',
+  appVersion: '2019-10-18',
+  name: 'giving',
+  paths,
+};