+++ src/http-client/index.ts
@@ -1,1 +1,2 @@
 export * from './http-client.js';
+export * from './types.js';