+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from './pco-api-client.js';
+export * from './pco-api-client/pco-api-client.js';