+++ src/pco-api-client/pco-api-client-config.ts
@@ -1,0 +1,8 @@
+export interface PCOApiClientConfig {
+  auth: PCOApiClientAuthPAT;
+}
+
+export interface PCOApiClientAuthPAT {
+  secret: string;
+  token: string;
+}